Разделить main.js на модули ES2015
